requirements.txt
# 
# Tiramisu Discord Bot
# --------------------
# Dependencies
# 

nextcord
nextcord-ext-menus
logging42
pyyaml
mysql-connector-python
